+++ resip/stack/DnsResult.hxx
@@ -117,7 +117,7 @@ class DnsResult : public DnsResultSink
 
       // return the target of associated query
       Data target() const { return mTarget; }
-      unsigned int getSRVResultsSize() const {return mSRVResults.size();}
+      unsigned int getSRVResultsSize() const {return (unsigned int)mSRVResults.size();}
 
       // Will delete this DnsResult if no pending queries are out there or wait
       // until the pending queries get responses and then delete
